+++ bot/database/tests/guild_repository.py
@@ -35,7 +35,8 @@ CHANGED_PREFIX = DEFAULT_PREFIX + '..'
 class GuildRepositoryTest(Test):
     def __init__(self):
         self.session = Configs.session
-        self.repository = GuildRepository(self.session)
+        if self.session is not None:
+            self.repository = GuildRepository(self.session)
 
     async def test_01_guild_create(self):
         guild = Guild(GUILD_ID_TEST)
@@ -50,8 +51,9 @@ class GuildRepositoryTest(Test):
     async def test_03_guild_get(self):
         guild = await self.repository.find(GUILD_ID_TEST)
         self.assert_is_not_none(guild)
-        self.assert_equal(guild.id, GUILD_ID_TEST)
-        self.assert_equal(guild.prefix, DEFAULT_PREFIX)
+        if guild:
+            self.assert_equal(guild.id, GUILD_ID_TEST)
+            self.assert_equal(guild.prefix, DEFAULT_PREFIX)
 
     async def test_04_guild_get_fail(self):
         result = await self.repository.find(GUILD_ID_TEST + 2)
@@ -59,18 +61,24 @@ class GuildRepositoryTest(Test):
 
     async def test_05_guild_update(self):
         guild = await self.repository.find(GUILD_ID_TEST)
-        guild.prefix = CHANGED_PREFIX
-        await self.repository.update(guild)
+        self.assert_is_not_none(guild)
+        if guild:
+            guild.prefix = CHANGED_PREFIX
+            await self.repository.update(guild)
 
     async def test_06_guild_update_confirm(self):
         guild = await self.repository.find(GUILD_ID_TEST)
-        self.assert_equal(guild.id, GUILD_ID_TEST)
-        self.assert_equal(guild.prefix, CHANGED_PREFIX)
+        self.assert_is_not_none(guild)
+        if guild:
+            self.assert_equal(guild.id, GUILD_ID_TEST)
+            self.assert_equal(guild.prefix, CHANGED_PREFIX)
 
     async def test_07_guild_delete_by_id(self):
         guild = await self.repository.find(GUILD_ID_TEST)
-        result = await self.repository.delete(guild)
-        self.assert_true(result)
+        self.assert_is_not_none(guild)
+        if guild:
+            result = await self.repository.delete(guild)
+            self.assert_true(result)
 
     async def test_08_guild_delete_confirm(self):
         result = await self.repository.find(GUILD_ID_TEST)
